requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
jdcal==1.4.1
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
